+++ packages/plugin-packages/mashroom-http-proxy/type-definitions/internal.ts
@@ -9,6 +9,7 @@ export interface HttpHeaderFilter {
 }
 
 export type PoolConfig = {
+    keepAlive: boolean;
     rejectUnauthorized: boolean;
     maxSockets: number;
 }